Move test endpoint to separate tag. Fix Spring annotation parsing.
